+++ Source/MQTTnet.AspnetCore/Client/Tcp/TcpConnection.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.IO.Pipelines;
 using System.Net;
